Rename errors -> error codes. Change module name to error
