+++ js/bundles/modules/file_management.js
@@ -1,6 +1,6 @@
 /* eslint-disable import/no-commonjs */
 import DevExpress from './core';
-import FileSystemError from '../../file_management/file_system_error';
+import FileSystemError from '../../file_management/error';
 import FileSystemItem from '../../file_management/file_system_item';
 import ObjectFileSystemProvider from '../../file_management/object_provider';
 import RemoteFileSystemProvider from '../../file_management/remote_provider';
