+++ js/ui/file_manager/file_items_controller.js
@@ -3,8 +3,8 @@ import FileSystemItem from '../../file_management/file_system_item';
 import ObjectFileSystemProvider from '../../file_management/object_provider';
 import RemoteFileSystemProvider from '../../file_management/remote_provider';
 import CustomFileSystemProvider from '../../file_management/custom_provider';
-import FileSystemError from '../../file_management/file_system_error';
-import ErrorCode from '../../file_management/errors';
+import FileSystemError from '../../file_management/error';
+import ErrorCode from '../../file_management/error_codes';
 import { pathCombine, getEscapedFileName, getPathParts, getFileExtension } from '../../file_management/utils';
 import { whenSome } from './ui.file_manager.common';
