+++ js/ui/file_manager/ui.file_manager.messages.js
@@ -1,5 +1,5 @@
 import messageLocalization from '../../localization/message';
-import ErrorCode from '../../file_management/errors';
+import ErrorCode from '../../file_management/error_codes';
 
 export const FileManagerMessages = {
     get: (errorCode, args) => {
